requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2==2.9.4
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
